+++ R/count-tally.R
@@ -61,8 +61,12 @@
 #' # mutate() instead of summarise
 #' df %>% add_count(gender, wt = runs)
 #' df %>% add_tally(wt = runs)
-count <- function(x, ..., wt = NULL, sort = FALSE, name = NULL, .drop = group_by_drop_default(x)) {
+count <- function(x, ..., wt = NULL, sort = FALSE, name = NULL) {
+  UseMethod("count")
+}
 
+#' @export
+count.data.frame <- function(x, ..., wt = NULL, sort = FALSE, name = NULL, .drop = group_by_drop_default(x)) {
   if (!missing(...)) {
     out <- group_by(x, ..., .add = TRUE, .drop = .drop)
   } else {
@@ -78,9 +82,16 @@ count <- function(x, ..., wt = NULL, sort = FALSE, name = NULL, .drop = group_by
   out
 }
 
+count.tbl_sql <- count.data.frame
+
 #' @export
 #' @rdname count
 tally <- function(x, wt = NULL, sort = FALSE, name = NULL) {
+  UseMethod("tally")
+}
+
+#' @export
+tally.data.frame <- function(x, wt = NULL, sort = FALSE, name = NULL) {
   n <- tally_n(x, {{ wt }})
   name <- check_name(name, group_vars(x))
 
@@ -96,6 +107,8 @@ tally <- function(x, wt = NULL, sort = FALSE, name = NULL) {
     out
   }
 }
+
+tally.tbl_sql <- tally.data.frame
 
 #' @export
 #' @rdname count