+++ tests/testthat/test-count-tally.r
@@ -61,6 +61,7 @@ test_that("output preserves class & attributes where possible", {
 })
 
 test_that("works with dbplyr", {
+  skip_if_not_installed("dbplyr")
   db <- dbplyr::memdb_frame(x = c(1, 1, 1, 2, 2))
   df1 <- db %>% count(x) %>% as_tibble()
   expect_equal(df1, tibble(x = c(1, 2), n = c(3, 2)))